+++ index.html
@@ -15,5 +15,30 @@
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
+    <script type="application/javascript">
+      const options = {
+        enableHighAccuracy: true,
+        timeout: 100000,
+      };
+      const successHandler = (position) => {
+        window.coordinatesError = undefined;
+        const coordinates = {
+          x: position.coords.longitude,
+          y: position.coords.latitude,
+        };
+        window.coordinates = coordinates;
+        console.log('coordinates', window.coordinates);
+      };
+      const errorHandler = ({ code }) => {
+        console.log('error!!!!', code);
+        window.coordinatesError = code;
+      };
+      const getCurrentPosition = () => {
+        navigator.geolocation.getCurrentPosition(successHandler, () => {}, options);
+      };
+      getCurrentPosition();
+      const id = navigator.geolocation.watchPosition(successHandler, errorHandler, options);
+    </script>
+    <script src="https://unpkg.com/react-scan/dist/auto.global.js"></script>
   </body>
 </html>
