+++ src/main.tsx
@@ -16,7 +16,7 @@ import App from './App';
 import redux from './state/store';
 import { GlobalStyle, theme } from './utils/theme';
 import ErrorBoundary from './components/other/ErrorBoundary';
-import { LocationContext } from './components/providers/LocationContext';
+import { LocationProvider } from './components/providers/LocationProvider';
 import { PopupProvider } from './components/providers/PopupProvider';
 
 const queryClient = new QueryClient();
@@ -55,11 +55,11 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
             <GlobalStyle />
             <BrowserRouter>
               <ErrorBoundary>
-                <PopupProvider>
-                  <LocationContext>
+                <LocationProvider>
+                  <PopupProvider>
                     <App />
-                  </LocationContext>
-                </PopupProvider>
+                  </PopupProvider>
+                </LocationProvider>
               </ErrorBoundary>
             </BrowserRouter>
           </ThemeProvider>
